requirements.txt
streamlit>=1.32.0
pymupdf>=1.23.0
anthropic>=0.25.0
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-httplib2>=0.2.0
python-dotenv>=1.0.0
